+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [4.45.1]
 * Rebasing Hotfixes:
+* Fixed Custom Domain Help Links
+* Fixed inability to change a user's role
 * Fixed user statuses
 * Re-added externalId
 * Isolated Logi script